requirements.txt
diffusers==0.25.0
gradio==6.5.1
lpips==0.1.4
lunar_tools==0.4
lunar_tools==0.0.12
matplotlib==3.10.8
numpy==2.4.2
pandas==3.0.0
Pillow==12.1.0
python_docx==1.2.0
PyYAML==6.0.3
scipy==1.17.0
setuptools==80.9.0
torch==2.10.0
torch==2.5.1+cu121
tqdm==4.67.1
transformers==4.36.2
